documented all newly introduced public members
