--- src/SourceGeneratorUtils/Descriptors/SourceFileDescriptor.cs
@@ -1,3 +1,0 @@
-﻿namespace SourceGeneratorUtils;
-
-public readonly record struct SourceFileDescriptor(string Name, SourceWriter Content);
